Add files via upload
This is my solution for the practical work given to us in ALSDD during the season 2023/24, it is an implementation of 4 types of traversals on binary search trees in C and Z languages (Z is an academic language created by the professor Zeggour). In this repository you will find both statement and solution of the assignement.